+++ index.ts
@@ -1,0 +1,3 @@
+export {CombatReadyApi} from './src/module/api';
+export {CombatReadyAnimationTheme} from './src/module/themes';
+export {CombatReadyTimer} from './src/module/timers';